Hosted Brand colours - extract the logic which calculates the text contrast colour to a new module
