+++ src/capi-multiple-hosted/web/index.html
@@ -1,5 +1,5 @@
 <div class="creative creative--hosted gs-container with-[%numberOfElements%]-elements"
-     data-link-name="Hosted Thrasher Multi" data-brand-color="[%BrandColour%];">
+     data-link-name="Hosted Thrasher Multi" data-brand-color="[%BrandColour%]">
     <div class="fc-container__inner" style="
             border-color: [%BrandColour%]
         ">
