+++ src/hosted-native-traffic-driver/amp/index.html
@@ -2,7 +2,7 @@
       font-family="Guardian Text Sans Web"
       timeout="1000"></amp-font>
 
-<div class="creative creative--hosted-mpu dark-font">
+<div class="creative creative--hosted-mpu">
     <a href="%%CLICK_URL_UNESC%%%%DEST_URL%%" class="creative__ctu" rel="nofollow" target="_top">
             <amp-img layout="fixed-height" class="creative__image" src="[%MainImage%]" height="160"></amp-img>
         <div class="creative__content">
