+++ src/hosted-native-traffic-driver/test.json
@@ -1,7 +1,7 @@
 {
 
   "Brand": "OMGB Visit Britain",
-  "BrandColour": "rgb(128,0,128)",
+  "BrandColour": "#f499b8",
   "LogoImage": "https://static.theguardian.com/commercial/hosted/gallery-prototype/omgb.png",
   "MainImage": "https://static.theguardian.com/commercial/hosted/gallery-prototype/omgb4.jpg",
   "Title": "Don't be a sloth this Summer",
