+++ src/hosted-native-traffic-driver/web/index.html
@@ -1,4 +1,4 @@
-<div id="creative" class="creative creative--hosted-mpu dark-font" style="background-color: [%BrandColour%];">
+<div id="creative" class="creative creative--hosted-mpu" data-brand-color="[%BrandColour%]" style="background-color: [%BrandColour%];">
 
     <a href="%%CLICK_URL_UNESC%%%%DEST_URL%%" class="creative__ctu" rel="nofollow" target="_top">
         <div id="ImageContainer"></div>
